desafio1/demoqa_qa_challenge/requirements.txt
requests==2.32.3
pytest==8.2.2
pytest-html==4.1.1
python-dotenv==1.0.1
selenium==4.23.1
webdriver-manager==4.0.2
faker==26.0.0
pytest-bdd==7.1.2
